+++ src/any_agent/serving/__init__.py
@@ -1,5 +1,4 @@
 from .config_mcp import MCPServingConfig
-from .server_handle import ServerHandle
 from .server_mcp import (
     serve_mcp,
     serve_mcp_async,
@@ -7,7 +6,6 @@ from .server_mcp import (
 
 __all__ = [
     "MCPServingConfig",
-    "ServerHandle",
     "serve_mcp",
     "serve_mcp_async",
 ]
